Displayer now seems to have no major errors.

+++ simulation/src/entities/painters/Oval.java
@@ -26,19 +26,19 @@ public class Oval implements Painter {
 	
 	@Override
 	public void paint(Graphics g, int x, int y, int w, int h) {
-		int xCenter, yCenter;
-		xCenter = x + w / 2;
-		yCenter = y + h / 2;
+		double xCenter, yCenter;
+		xCenter = x + (double) w / 2;
+		yCenter = y + (double) h / 2;
 		int xReal, yReal, wReal, hReal;
 		if (verticalOrNot == true) {
 			yReal = y;
 			hReal = h;
-			xReal = xCenter - (int) (scale * w / 2);
+			xReal = (int) (xCenter - (scale * w / 2));
 			wReal = (int) (scale * w);
 		} else {
 			xReal = x;
 			wReal = w;
-			yReal = yCenter - (int) (scale * h / 2);
+			yReal = (int) (yCenter - (scale * h / 2));
 			hReal = (int) (scale * h);
 		}
 		g.fillOval(xReal, yReal, wReal, hReal);
