+++ simulation/src/entities/painters/Rectangle.java
@@ -24,12 +24,12 @@ public class Rectangle implements Painter {
 
 	@Override
 	public void paint(Graphics g, int x, int y, int w, int h) {
-		int xCenter, yCenter;
-		xCenter = x + w / 2;
-		yCenter = y + h / 2;
+		double xCenter, yCenter;
+		xCenter = x + (double) w / 2;
+		yCenter = y + (double) h / 2;
 		int xReal, yReal, wReal, hReal;
-		xReal = xCenter - (int) (scale * w / 2);
-		yReal = yCenter - (int) (scale * h / 2);
+		xReal = (int) (xCenter - (scale * w / 2));
+		yReal = (int) (yCenter - (scale * h / 2));
 		wReal = (int) (scale * w);
 		hReal = (int) (scale * h);
 		g.fillRect(xReal, yReal, wReal, hReal);
